openai-lambda/requirements.txt
attrs==25.3.0
aws-cdk-lib==2.181.1
constructs==10.4.2
boto3==1.37.18
botocore==1.37.18
cattrs==24.1.2
exceptiongroup==1.2.2
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.110.0
publication==0.0.3
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==1.26.20
zipp==3.21.0
